Rename omaha to plo4
