+++ cpp/examples/.gitignore
@@ -1,5 +1,5 @@
 c_example
 cpp_example
-omaha_example
+plo4_example
 plo5_example
 plo6_example
